refactor: rename @/src/components/city-weather.tsx to @/src/components/CityWeather.tsx
